+++ internal/interp/stmt.go
@@ -25,6 +25,10 @@ func (i *Interpreter) execStmt(stmt ast.Stmt) (ctrl control, err error) {
 	}
 	defer func() { err = failure(stmt.Start(), diag.RType, err) }()
 	switch s := stmt.(type) {
+	case *ast.EchoStmt:
+		return control{}, i.execEcho(s)
+	case *ast.ChronometerStmt:
+		return control{}, i.execChronometer(s)
 	case *ast.RandomInputStmt:
 		return control{}, i.execRandomInput(s)
 	case *ast.AssignStmt: